third_party/WebKit/LayoutTests/animations/animation-pseudo-rem.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
:root {
  font-size: 1234px;
}
@keyframes remAnimation {
  from {
    width: 10rem;
  }
  to {
    width: 10rem;
  }
}
#element:before {
  content: '';
  display: block;
  animation: remAnimation 1s forwards;
}
</style>
<div id='element'></div>

<script>
test(function() {
  assert_equals(getComputedStyle(element, ':before').width, '12340px');
}, 'Check rem is calculated in animation keyframe');
</script>
